Moved bokeh plot baseclasses into own module
